+++ bio_utils/bio_programs/count_reads.py
@@ -6,6 +6,7 @@ import pandas as pd
 
 import bio_utils.fastx_utils as fastx_utils
 import misc.utils as utils
+import misc.pandas_utils as pandas_utils
 
 import logging
 import misc.logging_utils as logging_utils
@@ -44,7 +45,7 @@ def main():
     msg = "Writing data frame to disk"
     logger.info(msg)
 
-    utils.write_df(df, args.out, index=False)
+    pandas_utils.write_df(df, args.out, index=False)
 
 
 if __name__ == '__main__':